ip_tracking/requirements.txt
Django>=4.2
django-ratelimit>=4.0.0
ip2geotools>=0.1.6
celery>=5.3
redis>=5.0.0
django-celery-beat>=2.5.0
